feat(dashboard): refactor variables input
