+++ apps/api-gql/schema/games.graphqls
@@ -86,7 +86,7 @@ type SeppukuGame {
 
 input SeppukuGameOpts {
 	enabled: Boolean
-	timeoutSeconds: Int @validate(constraint: "max=84600,min=1")
+	timeoutSeconds: Int @validate(constraint: "max=86400,min=1")
 	timeoutModerators: Boolean
 	message: String @validate(constraint: "max=500")
 	messageModerators: String @validate(constraint: "max=500")
@@ -115,7 +115,7 @@ type VotebanGame {
 
 input VotebanGameOpts {
 	enabled: Boolean
-	timeoutSeconds: Int @validate(constraint: "min=1,max=84600")
+	timeoutSeconds: Int @validate(constraint: "min=1,max=86400")
 	timeoutModerators: Boolean
 	initMessage: String @validate(constraint: "max=500")
 	banMessage: String @validate(constraint: "max=500")
@@ -123,7 +123,7 @@ input VotebanGameOpts {
 	surviveMessage: String @validate(constraint: "max=500")
 	surviveMessageModerators: String @validate(constraint: "max=500")
 	neededVotes: Int @validate(constraint: "min=1,max=999999")
-	voteDuration: Int @validate(constraint: "min=1,max=84600")
+	voteDuration: Int @validate(constraint: "min=1,max=86400")
 	votingMode: VoteBanGameVotingMode @validate(constraint: "max=500")
 	chatVotesWordsPositive: [String!] @validate(constraint: "dive,max=500")
 	chatVotesWordsNegative: [String!] @validate(constraint: "dive,max=500")
